feat(page-title): update marquee helper to be mb string compatible
